Tip input incorporated. Requires tip to be added to total.

+++ CheckPlz/app/src/main/java/com/example/h/checkplz/PersonBill.java
@@ -71,7 +71,7 @@ public class PersonBill implements Parcelable {
         return mTip;
     }
 
-    public void setmTip(int mTip) {
+    public void setmTip(double mTip) {
         this.mTip = mTip;
     }
